+++ client/src/addirsync.rs
@@ -1,6 +1,7 @@
 use crate::LdapClient;
 use crate::*;
 use base64urlsafedata::Base64UrlSafeData;
+use ldap3_proto::control::LdapControl;
 
 #[derive(Debug)]
 pub struct LdapSyncReplEntry {
